Multiball Tweaks
Changed the key for multiball in input to b to avoid toggling mute and
triggering multiball at the same time. And now that I have a fresh set
of ears today I don't mind the sound effect I made for the multiball
yesterday. But as always I'm open to suggestions. I just made it a bit
louder in reaper (the DAW) for better balancing in game. PS - I'm
practicing committing changes from the command line, so if this gets
messed up, I'm sorry and please let me know!